src/pyproject.toml
[tool.poetry]
name = "generative-ai-ec-demo"
version = "0.1.0"
description = ""
authors = ["horiuyas <horiuyas@amazon.co.jp>", "nakayukl <nakayukl@amazon.co.jp>", "kobayasd <kobayasd@amazon.co.jp>", "daikiszk <daikiszk@amazon.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
opencv-python = "^4.9.0.80"
rembg = "^2.0.53"
boto3 = "^1.34.29"
matplotlib = "^3.8.2"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
faiss-cpu = "^1.8.0"
st-pages = "1.0.1"
langchain = "0.2.9"
langchain-community = "0.2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
